Made expert work
Swapped x and y in getCell and replaceCell in Board.hs and changed the
gettting of the input coordinates in Game.hs so the 1st int is x and 2nd
is y and now expert works. I don't know if this fucks with other stuff.
It can be reverted if it does.